QnA DAO 기능 목록 작성

+++ src/main/java/com/toyProject/domain/AlbumVO.java
@@ -17,7 +17,7 @@ import lombok.ToString;
 @Builder
 public class AlbumVO {
 
-	private int ano;
+	private Long ano;
 	private String title;
 	private String content;
 	private String imageFileName;
